small code fixes
